backend/requirements.txt
fastapi
uvicorn
whisperx
torch
torchaudio
torchvision
sentence-transformers 
transformers 
chromadb
python-multipart 
ffmpeg-python 
pydub 
gtts
requests  
python-dotenv 
openai
docling
flask
imageio[ffmpeg]